requirements.txt
numpy
matplotlib
pyyaml
pygame
pyclothoids